db restructuring begins
